refactor(desktop): orchestrator 预编排任务按「一任务一文件」拆到 flows/
把有状态协调器里的预编排任务从 god-class 方法拆成 flows/ 下的自由函数，经 OrchestratorRuntime 复用
协调器的运行态与共享 helper（不各自持有可变状态）：

- 新增 runtime.ts：AgentChat 类型 + OrchestratorRuntime 接口（registerController/clearController/
  markRunning/unmarkRunning/emitStep/takePending/withAgentChat/recordReviewSummaryMessage/setAutopilotBusy
  + ctx/runQueue）。
- flows/review.ts：reviewFlow（手动评审编排）+ runReviewForPr（组装 ReviewDeps + 调 runner）。
- flows/planning.ts：planningFlow（自由规划编排）+ runPlanningForPr。
- flows/autopilot.ts：autopilotPass（AutoPilot 批量准入 + 并行编排）。
- orchestrator.ts 退为瘦协调器：实例可变状态 + 公共 API（委托各 flow）+ runtime helper 实现，
  implements OrchestratorRuntime（把 this 作 runtime 传入 flow）。参数名统一用 runtime（不用 rt）。

纯搬运 / 拆分、行为不变；desktop typecheck/lint/build 全绿。
